code cleanup

+++ uninstall.sh
@@ -4,14 +4,16 @@
 # a `*boot` repo.
 
 # The hostname is in /etc/hostname prior to running `install.sh` here!
-HOSTNAME=$(cat /etc/hostname)
+HOSTNAME=$(hostname)
 
 echo -n "Started UNinstalling domog on "; date
 
 pushd "${HOME}/domog"
- source ./includes
+  source ./includes
 
-  sudo rm /etc/cron.d/domog
+  if [ -e /etc/cron.d/domog ]; then
+    sudo rm /etc/cron.d/domog
+  fi
 
   echo "  Stopping all service daemons"
   for daemon in $srvclist; do
